AI/requirements.txt
tensorflow==2.13.0rc0
Pillow==9.5.0
numpy==1.23.5
diffusers
transformers
imutils
torch==1.9.0
fastapi
uvicorn
accelerate
opencv-python==4.7.0.72
dlib==19.24.1